* origin/develop:
  fix a small problem with vocab_service
  more details on orcid_widget display, internal documentation
  updated README
  added readme file, fix small bugs, updated documentation
  documentation for orcid widget
  various fixes for orcid widget
  first cut at refactoring orcid widget
  updated readme

+++ vocab_service/models/vocab_services.php
@@ -265,29 +265,31 @@ class Vocab_services extends CI_Model {
 
 	function undoVersions($vocab_id){
 		$versions = $this->getVersionsByID($vocab_id);
-		foreach($versions as $version){
+		if(is_array($versions) && sizeof($versions) > 0){
+			foreach($versions as $version){
 
-			//make all current superseded
-			if($version->status=='current'){
-				$data = array(
-					'status'=>'superseded'
-				);
-				$this->vocab_db->where('id', $version->id);
-				$this->vocab_db->update('vocab_versions', $data);
-			}
+				//make all current superseded
+				if($version->status=='current'){
+					$data = array(
+						'status'=>'superseded'
+					);
+					$this->vocab_db->where('id', $version->id);
+					$this->vocab_db->update('vocab_versions', $data);
+				}
 
-			//supersede pending delete
-			if($version->status=="pending-delete"){
-				$data = array(
-					'status'=>'superseded'
-				);
-				$this->vocab_db->where('id', $version->id);
-				$this->vocab_db->update('vocab_versions', $data);
-			}
+				//supersede pending delete
+				if($version->status=="pending-delete"){
+					$data = array(
+						'status'=>'superseded'
+					);
+					$this->vocab_db->where('id', $version->id);
+					$this->vocab_db->update('vocab_versions', $data);
+				}
 
-			//delete all pending add
-			if($version->status=="pending-add"){
-				$this->vocab_db->delete('vocab_versions', array('id' => $version->id)); 
+				//delete all pending add
+				if($version->status=="pending-add"){
+					$this->vocab_db->delete('vocab_versions', array('id' => $version->id)); 
+				}
 			}
 		}
